requirements.txt
streamlit
pandas
scikit-learn
pytesseract
pillow
pyttsx3

